@@ -723,9 +723,6 @@ func (c *compiler) compileTry(e *Try) error {
 func (c *compiler) compileReduce(e *Reduce) error {
 	c.appendCodeInfo(e)
 	defer c.newScopeDepth()()
-	setfork := c.lazy(func() *code {
-		return &code{op: opfork, v: len(c.codes)}
-	})
 	c.append(&code{op: opdup})
 	v := c.newVariable()
 	f := c.newScopeDepth()
@@ -734,6 +731,9 @@ func (c *compiler) compileReduce(e *Reduce) error {
 	}
 	f()
 	c.append(&code{op: opstore, v: v})
+	setfork := c.lazy(func() *code {
+		return &code{op: opfork, v: len(c.codes)}
+	})
 	if err := c.compileQuery(e.Query); err != nil {
 		return err
 	}
@@ -910,8 +910,8 @@ func (c *compiler) compileFunc(e *Func) error {
 			env := make(map[string]any)
 			if c.environLoader != nil {
 				for _, kv := range c.environLoader() {
-					if i := strings.IndexByte(kv, '='); i > 0 {
-						env[kv[:i]] = kv[i+1:]
+					if k, v, ok := strings.Cut(kv, "="); ok && k != "" {
+						env[k] = v
 					}
 				}
 			}
@@ -934,20 +934,24 @@ func (c *compiler) compileFunc(e *Func) error {
 		return c.compileCallPc(f, e.Args)
 	}
 	if fds, ok := builtinFuncDefs[e.Name]; ok {
+		var compiled bool
 		for _, fd := range fds {
 			if len(fd.Args) == len(e.Args) {
 				if err := c.compileFuncDef(fd, true); err != nil {
 					return err
 				}
+				compiled = true
 				break
 			}
 		}
-		if len(fds) == 0 {
+		if !compiled {
 			switch e.Name {
 			case "_assign":
 				c.compileAssign()
 			case "_modify":
 				c.compileModify()
+			case "last":
+				c.compileLast()
 			}
 		}
 		if f := c.lookupBuiltin(e.Name, len(e.Args)); f != nil {
@@ -1059,13 +1063,13 @@ func (c *compiler) compileAssign() {
 		&code{op: opfork, v: len(c.codes) + 30}, // reduce [L1]
 		&code{op: opdup},
 		&code{op: opstore, v: w},
-		&code{op: oppathbegin}, // path(p)
+		&code{op: oppathbegin}, //                  path(p)
 		&code{op: opload, v: p},
 		&code{op: opcallpc},
 		&code{op: opload, v: w},
 		&code{op: oppathend},
 		&code{op: opstore, v: q}, //                as $q (.;
-		&code{op: opload, v: a},  //                setpath($q; $x)
+		&code{op: opload, v: a},  //                  setpath($q; $x)
 		&code{op: opload, v: x},
 		&code{op: opload, v: q},
 		&code{op: opload, v: w},
@@ -1131,6 +1135,35 @@ func (c *compiler) compileModify() {
 		&code{op: opload, v: d},
 		&code{op: opload, v: v},
 		&code{op: opcall, v: [3]any{funcDelpathsWithAllocator, 2, "_delpaths"}},
+		&code{op: opret},
+	)
+}
+
+// Appends the compiled code for the `last/1` function to
+// maximize performance avoiding unnecessary boxing.
+func (c *compiler) compileLast() {
+	defer c.appendBuiltin("last", 1)()
+	scope := c.newScope()
+	v, g, x := [2]int{scope.id, 0}, [2]int{scope.id, 1}, [2]int{scope.id, 2}
+	c.appends(
+		&code{op: opscope, v: [3]int{scope.id, 3, 1}},
+		&code{op: opstore, v: v},
+		&code{op: opstore, v: g},
+		&code{op: oppush, v: true}, //              $x = true
+		&code{op: opstore, v: x},
+		&code{op: opload, v: v},
+		&code{op: opfork, v: len(c.codes) + 13}, // reduce [L1]
+		&code{op: opload, v: g},                 // g
+		&code{op: opcallpc},
+		&code{op: opstore, v: v},    //             as $v (
+		&code{op: oppush, v: false}, //               $x = false
+		&code{op: opstore, v: x},
+		&code{op: opbacktrack},  //                 );
+		&code{op: oppop},        //                 [L1]
+		&code{op: opload, v: x}, //                 if $x then $v else empty end
+		&code{op: opjumpifnot, v: len(c.codes) + 17},
+		&code{op: opbacktrack},
+		&code{op: opload, v: v},
 		&code{op: opret},
 	)
 }
